attention-monitor/requirements.txt
opencv-python
mediapipe
numpy
pandas
matplotlib
requests
flask
fastapi
uvicorn[standard]
websockets
pydantic
pytest
python-dotenv
slowapi
